Fixed all eslint errors
