+++ package.json
@@ -12,10 +12,10 @@
     "styled-components": "^2.1.1"
   },
   "devDependencies": {
-    "eslint": "^4.2.0",
+    "eslint": "^3.19.0",
     "eslint-config-airbnb": "^15.0.2",
     "eslint-plugin-import": "^2.7.0",
-    "eslint-plugin-jsx-a11y": "^6.0.2",
+    "eslint-plugin-jsx-a11y": "^5.1.1",
     "eslint-plugin-react": "^7.1.0",
     "react-scripts": "1.0.10"
   },
