+++ src/App.js
@@ -1,10 +1,10 @@
 import React, { Component } from 'react';
-import { BrowserRouter as Router, Route } from 'react-router-dom'
+import { BrowserRouter as Router, Route } from 'react-router-dom';
 import styled from 'styled-components';
 
 import './App.css';
-import Complexes from './Complexes/List'
-import Complex from './Complexes/View'
+import Complexes from './Complexes/List';
+import Complex from './Complexes/View';
 import Header from './Header';
 import Footer from './Footer';
 
@@ -16,5 +16,4 @@ export default props =>
       <Route exact path="/сomplex" component={Complex} />
       <Footer />
     </div>
-  </Router>
-;
+  </Router>;
