+++ src/Complexes/List/index.jsx
@@ -1,15 +1,14 @@
-import React, { Component } from "react";
-import styled from "styled-components";
-import { Link } from "react-router-dom";
-import { Grid, Row, Col } from "react-flexbox-grid";
-import Hero from "./Hero";
-import Headline from "./Headline";
-import Card from "./Card";
+import React from 'react';
+import styled from 'styled-components';
+import { Grid } from 'react-flexbox-grid';
+import Hero from './Hero';
+import Headline from './Headline';
+import Card from './Card';
 
 const Cards = styled.div`padding-bottom: 3rem;`;
 
-export default props =>
-  <div>
+export default() =>
+  (<div>
     <Hero />
     <Headline />
     <Grid>
@@ -20,12 +19,12 @@ export default props =>
         >
           The Lewis Steel Building is a masterful industrial conversion located
           in the heart of Williamsburg. Located at 76 North 4th Street, the
-          former 1930's steel factory has been transformed into 83 individually
+          former 1930&apos;s steel factory has been transformed into 83 individually
           unique and luxury loft apartments.
         </Card>
         <Card location="MIDTOWN EAST, MANHATTAN" title="100 East 53rd Street">
           One Hundred East Fifty Third Street by Foster + Partners is a limited
-          collection of modern residences in Midtown Manhattan's Cultural
+          collection of modern residences in Midtown Manhattan&apos;s Cultural
           District. The 94 residences ranging from alcove lofts to four bedrooms
           within the 63-story tower are generously proportioned.
         </Card>
@@ -38,4 +37,4 @@ export default props =>
         </Card>
       </Cards>
     </Grid>
-  </div>;
+  </div>);
