+++ src/Complexes/View/index.jsx
@@ -1,17 +1,17 @@
-import React, { Component } from "react";
-import BodyClassName from "react-body-classname";
-import styled from "styled-components";
-import { Grid, Row, Col } from "react-flexbox-grid";
+import React from 'react';
+import BodyClassName from 'react-body-classname';
+import styled from 'styled-components';
+import { Grid, Row, Col } from 'react-flexbox-grid';
 
-import Info from "./Info";
-import Photos from "./Photos";
-import Summary from "./Summary";
-import Special from "./Special";
-import Description from "./Description";
-import Infrastructure from "./Infrastructure";
-import Offer from "./Offer";
-import Guide from "./Guide";
-import Maps from "./Map";
+import Info from './Info';
+import Photos from './Photos';
+import Summary from './Summary';
+import Special from './Special';
+import Description from './Description';
+import Infrastructure from './Infrastructure';
+import Offer from './Offer';
+import Guide from './Guide';
+import Maps from './Map';
 
 const Offers = styled.div`
   padding: 4rem 0;
@@ -25,10 +25,10 @@ const Title = styled.h2`
   padding-bottom: 1.5rem;
 `;
 
-const complexName = "Жилой комплекс «Полянка/44»";
+const complexName = 'Жилой комплекс «Полянка/44»';
 
-export default props =>
-  <BodyClassName className="bg-white">
+export default() =>
+  (<BodyClassName className="bg-white">
     <div>
       <Info
         title={complexName}
@@ -76,4 +76,4 @@ export default props =>
       <Guide />
       <Maps />
     </div>
-  </BodyClassName>;
+  </BodyClassName>);
